+++ ChangeLog.md
@@ -28,6 +28,7 @@ Bug fixes:
 * Extensions from the `other-extensions` field no longer enabled by default [#449](https://github.com/commercialhaskell/stack/issues/449)
 * Fix: haddock forces rebuild of empty packages [#452](https://github.com/commercialhaskell/stack/issues/452)
 * Don't copy over executables excluded by component selection [#605](https://github.com/commercialhaskell/stack/issues/605)
+* Fix: stack fails on Windows with git package in stack.yaml and no git binary on path [#712](https://github.com/commercialhaskell/stack/issues/712)
 
 ## 0.1.2.0